Scripts: Add a script to generate a registration code
The registration code algorithm is now (mostly) reverse-engineered. The
process is dependent on the gKey value from the game source, which
should evaluate to 0x1E42A71F.

Registration requires a name of at least 4 letters. The name is
converted to all uppercase and spaces are removed.

The registration key is created in two parts. First, the last 4 letters
of the name are converted to an ascii hex representation and cast to an
int. For example the name "FREE" is 0x46524545. This is the name number.

Second, the first four letters of the name are converted to a list of
ascii hex values and summed. The result mod 0xFF is then used as a seed.
This seed XORed with the gKey XORed with the name number is what is used
for the first 8 characters of the code.

The last two characters are the hex representation of the seed. These
two strings are joined for the final registration code.