Quickdraw experiments
Added tests/quickdraw_test.c and related files to implement a small
subset of the QuickDraw format. This code allows saving PPic 1000
through PPic 1008 to disk as a ppm file. PPic 1009 will require more
work in the QuickDraw library to implement properly.

This is a great verification that the decompression of packed data works
properly, as the images for the menu, help pages, and other screens are
now viewable extracted from the game binary!

For this to work properly, it was determined that lzrw3-a was the
correct algorithm, not lzrw3. That file has been replaced.