Sprites Test
Added a sprites test for experimenting with extracting sprites

The correct sprites are extracted into ppm files, but there are
no colors. There is only enough included information for 1 color per
pixel.

At this point it becomes hard to test the functions in separate files.
There are so many globals that things like drawing sprites depend
on so many other factors. From here on out it seems the best path
forward will be actually making the game code itself work.

Here we go!