Replace packs.o with a header file
My understanding is that it is difficult to create a binary file for
linking in a cross-platform way. Rather than using a packs.o binary, use
a header file generated by xxd to store the game data. Though this is
simpler, it does mean the data is stored in the repository in a 35M
file.