Style: Reorganize repository structure
In the process of moving to CMake, I moved some files to the src/
directory while leaving others in the original source/ and headers/
dirs. Because I explicitly tell CMake which files to use, it is safe to
move all files to the src/ and src/include directories.

This also removes some old test files now that real test cases have been
added.