Tests: Add level decryption test
Adds a test that verifies that level decryption works properly. The
original source only ever supported checking level 4, and it seems that
is safe enough for now. Maybe in the future it would be good to
determine the correct "Chck" values for levels 5-10, but it is probably
safe enough without additional tests.