Feat: Add preference file reading
This adds a simple parser for reading an INI-like preferences file. The
file is expected at ~/.config/open-reckless-drivin/prefs.ini, but that
path is subject to change in this early state.

The parser handles booleans, ints, and unquoted strings. Errors are
caught but not reported. Lines that fail to parse are ignored. Lines
starting with a `#` are comments. Blank lines and trailing whitespace
are ignored.

TODO:
* Add a test for preference file reading
* Create default preference file if it doesn't exist
* Load registration information from preference file