Tests: Add preference file reading test
This adds a test file and an example .ini file for stress-testing the
preferences reader. This test reads a set of expected values, then
ignores a set of invalid preferences.

It would be better in the future to also return errors from the
preference reader so they can be properly reported and handled. For now
this seems to work well, and the test even helped find a few bugs.