Feat: Registration checking
Now that preferences can be read from a file, read the registration name
and code and use to check if a valid name has been registered.

Because no work has been done on the UI, the early versions of Open
Reckless Drivin' will do configuration entirely from this preferences
file. Only in later versions will a UI preference editor be a priority.