feat: add resource loading and tests
Ports the resource.c file to Zig for loading of the Resource Fork data.
The main difference here is that we need to be a bit more specific on
the alignment of the data, and we use an iterator function rather than
pointer comparisons.

A doc comment has also been added documenting each resource type, id,
and description.

It's nice having tests built-in to the source file!