feat: add lzrw decompression and tests
Adds the LZRW3-A decompression algorithm as a linked C library. Also
adds tests to main to test this algorithm.

When decompressing the Resource data, the first 4 bytes are to be
ignored. This will likely be done at the callsite rather than in the
lzrw decompress function. Or maybe a public decompressResource function
can be used that strips the first four bytes.