feat: add resource dump utility
Adds a new subcommand dump-resource that can be used to write the bytes
of any resource to stdout. This is useful for inspecting the bytes while
development and debugging. Not sure if this will be in the final game or
not.

Run with `$ reckless-drivin dump-resource [type] [id] | xxd | less` or
similar to dump a resource to a pager to view the bytes.