add initial level loading logic
Adds the code to load level data from a pack. This is incomplete, and
leaks memory.

Fixes bugs in the reader, and updates the pack loading code to be more
generic. Rather than specifying getEntrySparse and getEntrySequential,
the getEntry function takes a struct and uses the correct private
function for the respective pack type.