refactor: specify byte alignment for packs
Add annotations to the pack types and return types to align to a u32.
This makes sure the pack data is aligned correctly.

I still haven't analyzed things closely, but this assumes that the
entries inside a pack are also aligned. I believe they are, but I can't
be positive without inspecting all of them.