refactor: use a linked list for objects
The original game uses a linked list for objects. This makes sense, as
there are few (no?) cases where indexing is required, but there are many
cases where iteration over all objects, insertions, and removals are
needed. A linked list will make these operations a bit cheaper, even if
the memory is fragmented a bit.