render: road and background textures
The roads render now!! Wooooo!

This changes the SDL texture from argb8888 to rgb555 to match the
internal structure used in the original game. (also saves memory!).

There are a LOT of @intToFloat and @floatToInt calls. Usually they are
fine, but occasionally they can cause problems if the int is out of
bounds. This was handled implicitly in C, but in Zig it isn't.