Add particle drawing code
Most of the particle drawing code is here now. The g16BitClut is only
used for drawing particle effects. This resource doesn't seem to be
bundled with the resources Jonas uploaded to GitHub.

So it seems I will need to open up the game on a Macintosh emulator and
open ResEdit to attempt to find these. Or maybe there is a default CLUT?
Not sure.